Implement indented JSON output
There still is some bothersome code duplication between the stream
encoder and the object encoder - This is apparent as there now exist
two macros to set up the context for serializing an aggregate,
WITH-AGGREGATE/OBJECT and WITH-AGGREGATE/STREAM.  The two could
probably be integrated by making the object aggregator use the stream
aggregator underneath.

Indentation is currently only available when serializing to a
json-output-stream, which is always used with the stream encoder.