Correctly skip whitespce when no input is available.
LISTEN reports false on a stream that is not at EOF, but that doesn't
have data available yet. As-is, reading from a network stream where
the next byte of data isn't available will cause SKIP-WHITESPACE to
stop skipping, which may cause the caller to read a whitespace
character later and become confused.

Rather than checking for available input with LISTEN, instead use
PEEK-CHAR's facility to return a distinguished EOF value and check for
that. When the next byte of data has not been transmitted yet,
PEEK-CHAR will block until it's available and continue skipping
whitespace if appropriate. EOF is still left for the caller to
consume.